Reload DB when it has changed underneath.
During transactions, there is a possibility that the underlying DB
might have changed state, either because new records have been
added/deleted, or if the DB has been (re)packed. `zsdb_reload()`
reloads the DB when that happens. This is an expensive operattion and
there is a lot of scope to improve the logic for reloading DB.