file: Silence assign+compare warnings
I don't quite understand if the compiler if saying that this is actually
ambiguous, or if its just often confused/misread by humans. Either way
it seems benign to just add the parens.

zeroskip-file.c:496:23: warning: using the result of an assignment as a condition without parentheses [-Wparentheses]
        if (st.st_uid |= f->st.st_uid)
            ~~~~~~~~~~^~~~~~~~~~~~~~~
zeroskip-file.c:496:23: note: place parentheses around the assignment to silence this warning
        if (st.st_uid |= f->st.st_uid)
                      ^
            (                        )
zeroskip-file.c:496:23: note: use '!=' to turn this compound assignment into an inequality comparison
        if (st.st_uid |= f->st.st_uid)
                      ^~
                      !=