explicitly zero struct stat
A { 0 } initializer will not zero composite types, which somehow st_mode
is on macOS (or at least the compiler believes it to be so; in the
headers it appears to be typedef'd to an internal __darwin_uint16_t).

GCC has an extension { } that will cope with this case, but its not
valid C99 and emits a different warning in Clang.

To sidestep all of this, just memset them explicitly.

mappedfile.c:362:33: warning: missing field 'st_mode' initializer
      [-Wmissing-field-initializers]
        struct stat stbuf = { 0 };