zeroskip: define _DARWIN_C_SOURCE on Darwin to get memset
memset was never standardised, so there's no standard way to enable it
(if it even exists). On Darwin, it's in string.h and needs
_DARWIN_C_SOURCE to be defined (and other things, like _ANSI_SOURCE, to
not be defined).

But, this is sort of a crappy way to do it. Its at the top because
cstring.h brings in string.h. Really this needs to be dealt with in
configure, but its not quite set up for that yet and its more than I
want to dive into right now.

Also, while hunting around, I came across:

  https://moarvm.github.io/coverage/libmoar/coverage/home/travis/build/MoarVM/MoarVM/src/platform/memmem.h.html

Which suggests Darwin's memmem isn't that great anyway, so maybe this is
better dealt with by just bundling a decent implementation.