Re-initialise `active` flag when unlocking.
When updating `.zsdb` files, a lock is acquired and the lock is
released once the updates are done. The `struct file_lock` interface
takes care of the locking mechanisms. We had missed out resetting the
`active` flag in `struct file_lock` to `0` when unlocking. This
resulted in really vague errors, and corrupt Zeroskip DB's.